Added checkbox to 'Get Password From Vault' step to allow secure component process properties
