+++ .gitignore
@@ -1,0 +1,22 @@
+# ignore all dotfiles by except git specific ones
+.*
+!.gitignore
+!.git*
+
+# ignore all temporary files and ahp specific files
+*.bak
+*.bom
+*.swp
+*~
+plugin-build.xml
+
+# ignore compilation related directories in the root of the project
+/build
+/dist
+/lib
+/ecbin
+/bin/
+i18n-scraper.groovy
+gradlew
+gradlew.bat
+/gradle/
