+++ gradle.properties
@@ -7,7 +7,7 @@ pluginName=cyberark
 # The apiVersion property is used to set the @API_VERSION@ variable in the plugin.xml file and is used in
 # combination with the pluginVersion property to set the @RELEASE_VERSION@ variable in the info.xml file (apiVersion.pluginVersion).
 # It is also used in the distributable .zip file name as shown previously.
-apiVersion=1
+apiVersion=2
 
 # The pluginVersion property is used to set the specific build of the version, which is used in
 # combination with the apiVersion property to set the @RELEASE_VERSION@ variable in the info.xml file (apiVersion.pluginVersion).
@@ -15,12 +15,3 @@ apiVersion=1
 # For local builds, specify dev for the pluginVersion property. For releases, use the first eight characters of the
 # commit ID, ${GIT_COMMIT:0:8}, which is done on Bluemix when a commit is pushed to the GitHub project.
 pluginVersion=dev
-
-
-systemProp.http.proxyHost=hybrid-web.global.blackspider.com
-systemProp.http.proxyPort=80
-systemProp.http.nonProxyHosts=*.nonproxyrepos.com|localhost
-systemProp.https.proxyHost=hybrid-web.global.blackspider.com
-systemProp.https.proxyPort=80
-systemProp.https.nonProxyHosts=*.nonproxyrepos.com|localhost
-
