+++ src/main/zip/plugin.xml
@@ -41,6 +41,23 @@
                      description="AppID configured in CyberArk PVWA">
         </property-ui>
       </property>
+      <property name="isSecure">
+        <property-ui type="checkBox"
+                     label="Secure Properties"
+                     description="Check this box to create CyberArk/password, CyberArk/username,
+                        and CyberArk/address as component process properties instead of output 
+                        properties. The password property will be a secure value, which is not
+                        possible with output properties.">
+        </property-ui>
+      </property>
+      <property name="componentProcessId">
+        <property-ui type="textBox"
+                     label="Component Process ID"
+                     description="Property to reference the ID of the component process."
+                     hidden="true"
+                     default-value="${p:process.id}">
+        </property-ui>
+      </property>
     </properties>
     <post-processing>
       <![CDATA[
@@ -54,7 +71,7 @@
     </post-processing>
     <command program="${GROOVY_HOME}/bin/groovy">
       <arg value="-cp"/>
-      <arg path="classes:lib/groovy-plugin-utils-1.0.jar"/>
+      <arg path="classes:lib/groovy-plugin-utils.jar:lib/uDeployRestClient.jar"/>
       <arg file="clipasswordsdk.groovy"/>
       <arg file="${PLUGIN_INPUT_PROPS}"/>
       <arg file="${PLUGIN_OUTPUT_PROPS}"/>
@@ -108,7 +125,7 @@
     </post-processing>
     <command program="${GROOVY_HOME}/bin/groovy">
       <arg value="-cp"/>
-      <arg path="classes:lib/groovy-plugin-utils-1.0.jar"/>
+      <arg path="classes:lib/groovy-plugin-utils.jar"/>
       <arg file="conjur_auth.groovy"/>
       <arg file="${PLUGIN_INPUT_PROPS}"/>
       <arg file="${PLUGIN_OUTPUT_PROPS}"/>
@@ -162,7 +179,7 @@
     </post-processing>
     <command program="${GROOVY_HOME}/bin/groovy">
       <arg value="-cp"/>
-      <arg path="classes:lib/groovy-plugin-utils-1.0.jar"/>
+      <arg path="classes:lib/groovy-plugin-utils.jar"/>
       <arg file="conjur_getvar.groovy"/>
       <arg file="${PLUGIN_INPUT_PROPS}"/>
       <arg file="${PLUGIN_OUTPUT_PROPS}"/>
