+++ src/main/zip/upgrade.xml
@@ -4,11 +4,5 @@
         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
 
   <migrate to-version="2">
-    <migrate-command name="Hello World">
-      <migrate-properties>
-        <!-- Adding a new property always requires the 'default' attribute -->
-        <migrate-property name="name" default="World"/>
-      </migrate-properties>
-    </migrate-command>
   </migrate>
 </plugin-upgrade>
